fix(model): ERD vs Goモデル差分修正・job_titles.code削除
- vaccines: InventoryID *uint64 追加（DB inventory_id カラムが未マッピングだった）
- insurances: CoverageRate *float64/numeric(5,2) → int NOT NULL（DBはinteger NOT NULL）
- shift_entries: StartTime/EndTime string → *string gorm:"type:time"（DBはtime型・nullable）
- job_titles: Code フィールド削除（不要と判断）

波及修正:
- handler/insurance_{request,response,handler}: CoverageRate を *int → int に統一
- handler/pet_response: petInsuranceNested.CoverageRate *float64 → int
- handler/shift_{handler,response}: *string 対応（nil=未設定、空文字変換）
- service/shift_entry_service: CreateShiftEntryInput.StartTime/EndTime → *string
- handler/job_title_{request,handler}: Code フィールド削除
- service/job_title_service: UpdateJobTitleInput.Code と buildFields から削除
- service/insurance_service_test: CoverageRate テストデータを int に修正

マイグレーション:
- 008_drop_job_titles_code.sql: job_titles.code カラムをDROP

ドキュメント:
- ERD.md: job_titles・chief_complaint_categories に description カラムを追記

+++ backend/internal/handler/insurance_request.go
@@ -4,16 +4,16 @@ type createInsuranceRequest struct {
 	Name         string `json:"name"          binding:"required"`
 	IsActive     bool   `json:"is_active"`
 	Description  string `json:"description"`
-	CoverageRate *float64 `json:"coverage_rate"`
-	ContactPhone string   `json:"contact_phone"`
-	SortOrder    int      `json:"sort_order"`
+	CoverageRate *int   `json:"coverage_rate"`
+	ContactPhone string `json:"contact_phone"`
+	SortOrder    int    `json:"sort_order"`
 }
 
 type updateInsuranceRequest struct {
-	Name         string   `json:"name"`
-	IsActive     *bool    `json:"is_active"`
-	Description  string   `json:"description"`
-	CoverageRate *float64 `json:"coverage_rate"`
+	Name         string `json:"name"`
+	IsActive     *bool  `json:"is_active"`
+	Description  string `json:"description"`
+	CoverageRate *int   `json:"coverage_rate"`
 	ContactPhone string `json:"contact_phone"`
 	SortOrder    int    `json:"sort_order"`
 }
